+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.0.35
 - removed default value -1 on forward delay on wls_jms_queue
+- wls_server new attributelistenportenabled
 
 ## 1.0.34
 - Copydomain FMW apps_dir fix